Drop support for older `tls`
The `.cabal` file required `tls >= 2.1` when the `crypton` flag was enabled,
but allowed for older versions of `tls` without it -- but if you then actually
tried to build with older versions of `tls`, the build would fail. It has been
relatively easy to support older versions of `tls` until now, but it's getting
tricker now, so this just drops this entirely.